+++ ruby-lsp-ree/lib/ruby_lsp/ruby_lsp_ree/parsing/parsed_document.rb
@@ -7,14 +7,15 @@ class RubyLsp::Ree::ParsedDocument
 
   attr_reader :ast, :package_name, :class_node, :fn_node, :class_includes,
     :link_nodes, :values, :action_node, :dao_node, :filters,
-    :bean_node, :bean_methods, :mapper_node, :links_container_block_node
+    :bean_node, :bean_methods, :mapper_node, :links_container_block_node, :aggregate_node
 
   def initialize(ast)
     @ast = ast
   end
 
   def links_container_node
-    @fn_node || @action_node || @dao_node || @bean_node || @mapper_node
+    # TODO don't use separate node, use one field for all and additional type field: links_container_node_type
+    @fn_node || @action_node || @dao_node || @bean_node || @mapper_node || @aggregate_node
   end
 
   def includes_link_dsl?
@@ -55,35 +56,42 @@ class RubyLsp::Ree::ParsedDocument
     return unless class_node
 
     @fn_node ||= class_node.body.body.detect{ |node| node.name == :fn }
-    @links_container_block_node = @fn_node&.block
+    @links_container_block_node ||= @fn_node&.block
   end
 
   def parse_action_node
     return unless class_node
 
     @action_node ||= class_node.body.body.detect{ |node| node.name == :action }
-    @links_container_block_node = @action_node&.block
+    @links_container_block_node ||= @action_node&.block
   end
 
   def parse_dao_node
     return unless class_node
 
     @dao_node ||= class_node.body.body.detect{ |node| node.name == :dao }
-    @links_container_block_node = @dao_node&.block
+    @links_container_block_node ||= @dao_node&.block
   end
 
   def parse_bean_node
     return unless class_node
 
     @bean_node ||= class_node.body.body.detect{ |node| node.name == :bean }
-    @links_container_block_node = @bean_node&.block
+    @links_container_block_node ||= @bean_node&.block
   end
 
   def parse_mapper_node
     return unless class_node
 
     @mapper_node ||= class_node.body.body.detect{ |node| node.name == :mapper }
-    @links_container_block_node = @mapper_node&.block
+    @links_container_block_node ||= @mapper_node&.block
+  end
+
+  def parse_aggregate_node
+    return unless class_node
+
+    @aggregate_node ||= class_node.body.body.detect{ |node| node.name == :aggregate }
+    @links_container_block_node ||= @aggregate_node&.block
   end
 
   def parse_class_includes