+++ ruby-lsp-ree/lib/ruby_lsp/ruby_lsp_ree/parsing/parsed_document_builder.rb
@@ -48,6 +48,7 @@ class RubyLsp::Ree::ParsedDocumentBuilder
     document.parse_bean_node
     document.parse_dao_node
     document.parse_mapper_node
+    document.parse_aggregate_node
     document.parse_class_includes
     document.parse_links