+++ ruby-lsp-ree/lib/ruby_lsp/ruby_lsp_ree/ree_indexing_enhancement.rb
@@ -6,7 +6,7 @@ module RubyLsp
     class ReeIndexingEnhancement < RubyIndexer::Enhancement
       include RubyLsp::Ree::ReeLspUtils
 
-      REE_INDEXED_OBJECTS = [:fn, :enum, :action, :dao, :bean, :mapper]
+      REE_INDEXED_OBJECTS = [:fn, :enum, :action, :dao, :bean, :mapper, :aggregate]
 
       def on_call_node_enter(node)
         return unless @listener.current_owner